+++ docs/chapters/changelog.rst
@@ -9,7 +9,16 @@ Version 0.6.0
 
 Released 28 June 2021
 
+- Warning: This upgrade requires a Database Upgrade (which can be handled interactively by an admin from the Dashboards 
+  page in JupyterHub).
+- Run dashboard as viewer option (:ref:`spawn_as_viewer`) so each user clones a dashboard when viewing it and runs it as 
+  themself instead of accessing the creator's dashboard server directly.
+- :ref:`jupyter_startpath_regex` to restrict Jupyter Tree source dashboards to a specified regex, e.g. to ensure all 
+  dashboards reference a file on a shared path.
 - Remove show_source_servers and require_source_server functionality from DockerSpawner.
+
+If you are using DockerSpawner please only upgrade if you are not using the source spawner functionality (where 
+Docker images are copied from existing spawners to create new dashboard images).
 
 Version 0.5.7
 ~~~~~~~~~~~~~